pyproject.toml
[tool.poetry]
name = "terminal-data-service"
version = "0.1.0"
description = ""
authors = ["Bety <Elisabeta.Onoriu@endava.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = ">=2.0.42,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.1"
boto3 = "^1.40.13"
awscli-local = "^0.22.2"
pydantic-settings = "^2.10.1"
localstack = "^4.7.0"
aiosqlite = "^0.21.0"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.4"
pytest = "^8.4.1"
awscli-local = "^0.22.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
